+++ src/reading.rs
@@ -8,11 +8,8 @@ where
 {
     /// Read the temperature from the sensor.
     pub fn read_temperature(&mut self) -> Result<f32, Error<E>> {
-        let mut data = [0; 2];
-        self.i2c
-            .write_read(self.address, &[Register::TEMPERATURE], &mut data)
-            .map_err(Error::I2C)?;
-        Ok(convert_temp_from_register(data[0], data[1]))
+        let data = self.read_register_u16(Register::TEMPERATURE)?;
+        Ok(convert_temp_from_register(data.msb, data.lsb))
     }
 }
 
@@ -22,11 +19,8 @@ where
 {
     /// Read whether the one-shot measurement result is ready.
     fn one_shot_measurement_is_ready(&mut self) -> Result<bool, Error<E>> {
-        let mut data = [0; 2];
-        self.i2c
-            .write_read(self.address, &[Register::CONFIG], &mut data)
-            .map_err(Error::I2C)?;
-        Ok((data[1] & BitFlagsLow::ONE_SHOT) != 0)
+        let config = self.read_register_u16(Register::CONFIG)?;
+        Ok((config.msb & BitFlagsHigh::ONE_SHOT) != 0)
     }
 
     /// Perform a one-shot temperature measurement.
@@ -54,12 +48,10 @@ where
         {
             Err(nb::Error::WouldBlock)
         } else {
-            let mut data = [0; 2];
-            self.i2c
-                .write_read(self.address, &[Register::TEMPERATURE], &mut data)
-                .map_err(Error::I2C)
+            let data = self
+                .read_register_u16(Register::TEMPERATURE)
                 .map_err(nb::Error::Other)?;
-            let temp = convert_temp_from_register(data[0], data[1]);
+            let temp = convert_temp_from_register(data.msb, data.lsb);
             self.a_temperature_conversion_was_started = false;
             Ok(temp)
         }
@@ -81,12 +73,9 @@ where
     /// [AlertPolarity](enum.AlertPolarity.html).
     #[allow(clippy::wrong_self_convention)]
     pub fn is_comparator_mode_alert_active(&mut self) -> Result<bool, Error<E>> {
-        let mut data = [0; 2];
-        self.i2c
-            .write_read(self.address, &[Register::CONFIG], &mut data)
-            .map_err(Error::I2C)?;
-        let is_alert_polarity_high = (data[1] & BitFlagsLow::ALERT_POLARITY) != 0;
-        let alert_status = (data[0] & BitFlagsHigh::ALERT) != 0;
+        let config = self.read_register_u16(Register::CONFIG)?;
+        let is_alert_polarity_high = (config.msb & BitFlagsHigh::ALERT_POLARITY) != 0;
+        let alert_status = (config.lsb & BitFlagsLow::ALERT) != 0;
         Ok(is_alert_polarity_high == alert_status)
     }
 }