+++ tests/common/mod.rs
@@ -13,10 +13,10 @@ impl Register {
     pub const T_HIGH: u8 = 0x03;
 }
 
-pub struct BitFlagsLow;
+pub struct BitFlagsHigh;
 
 #[allow(unused)]
-impl BitFlagsLow {
+impl BitFlagsHigh {
     pub const SHUTDOWN: u8 = 0b0000_0001;
     pub const THERMOSTAT: u8 = 0b0000_0010;
     pub const ALERT_POLARITY: u8 = 0b0000_0100;
@@ -26,18 +26,18 @@ impl BitFlagsLow {
     pub const ONE_SHOT: u8 = 0b1000_0000;
 }
 
-pub struct BitFlagsHigh;
+pub struct BitFlagsLow;
 
 #[allow(unused)]
-impl BitFlagsHigh {
+impl BitFlagsLow {
     pub const EXTENDED_MODE: u8 = 0b0001_0000;
     pub const ALERT: u8 = 0b0010_0000;
     pub const CONV_RATE0: u8 = 0b0100_0000;
     pub const CONV_RATE1: u8 = 0b1000_0000;
 }
 
-pub const DEFAULT_CONFIG_MSB: u8 = BitFlagsHigh::CONV_RATE1 | BitFlagsHigh::ALERT;
-pub const DEFAULT_CONFIG_LSB: u8 = BitFlagsLow::RESOLUTION;
+pub const DEFAULT_CONFIG_MSB: u8 = BitFlagsHigh::RESOLUTION;
+pub const DEFAULT_CONFIG_LSB: u8 = BitFlagsLow::CONV_RATE1 | BitFlagsLow::ALERT;
 
 pub fn setup(expectations: &[I2cTransaction]) -> Tmp1x2<I2cMock, marker::mode::Continuous> {
     let i2c = I2cMock::new(expectations);